Fixed previous and new commited issues

+++ getMePayload.py
@@ -14,20 +14,17 @@ app = CTk()
 app.title("GET Me A Payload")
 app.set_appearance_mode("dark")
 app.geometry("1000x880")
-
+app.title("GetMePayload | Author: @PrashantShrestha")
 
 wrapper1 = LabelFrame(app)
 wrapper1.grid(row=0, column=0, padx=10, pady=10)
-# wrapper1.pack(side="left", fill="both", padx=10, pady=10)
 
 
 mycanvas = CTkCanvas(wrapper1, width=200, height=860)
 mycanvas.grid(row=0, column=0, padx=10, pady=10)
-# mycanvas.pack(side="left", fill="both", expand=True)
 
 yscrollbar = ttk.Scrollbar(wrapper1, orient="vertical", command=mycanvas.yview)
 yscrollbar.grid(row=0, column=1, padx=10, pady=10, sticky="ns")
-# yscrollbar.pack(side=RIGHT, fill=Y)
 
 mycanvas.configure(yscrollcommand=yscrollbar.set)
 mycanvas.bind('<Configure>', lambda e: mycanvas.configure(scrollregion = mycanvas.bbox('all')))
@@ -37,15 +34,31 @@ mycanvas.create_window((0, 0), window=myframe, anchor="nw")
 
 
 
-url =  "https://github.com/swisskyrepo/PayloadsAllTheThings"
-r = requests.get(url)
-soup = BeautifulSoup(r.content, 'html.parser')
+try:
+    if os.path.exists("source/homepage.html"):
+        with open("source/homepage.html", "r") as f:
+            html = f.read()
+            # print(html)
+            soup = BeautifulSoup(html, "html.parser")
+   
+    else:
+        url =  "https://github.com/swisskyrepo/PayloadsAllTheThings"
+        r = requests.get(url)
+        os.makedirs('source', exist_ok=True)
+        soup = BeautifulSoup(r.content, 'html.parser')
+        # print(soup)
+        with open('source/homepage.html', 'w', encoding='utf-8') as f_out:
+            f_out.write(str(soup))
+except:
+    print("[-] Error: Could not connect to github")
+    print('[-] Error: Requires internet connection for the first time.')
+    exit()
+
+
 
 # create a directory name source with OS
 os.makedirs('source', exist_ok=True)
 
-with open('source/homepage.html', 'w', encoding='utf-8') as f_out:
-    f_out.write(soup.prettify())
 
 ancs = soup.find_all('a')
 
@@ -69,21 +82,18 @@ for i in dic:
     globals () ['button_%s' % i] = CTkButton(myframe, text=k, command=lambda j=i:getPayload(j))
     dynamic_button = eval(['button_%s' % i][0])
     dynamic_button.grid(pady=10)
-    # button = CTkButton(myframe, text=i, command=lambda j=i:getPayload(j))
-    # button.grid(pady=10)
+
 
 wrapper2 = CTkFrame(app)
 wrapper2.grid(row=0, column=1, padx=10, pady=10)
-# wrapper2.pack(fill="both", side=RIGHT, padx=10, pady=10)
+
 
 
 mycanvas2 = CTkCanvas(wrapper2, width=640, height=850)
 mycanvas2.grid(row=0, column=0, padx=10, pady=10)
-# mycanvas2.pack(side=LEFT, fill=BOTH, expand=True)
 
 yscrollbar2 = ttk.Scrollbar(wrapper2, orient="vertical", command=mycanvas2.yview)
 yscrollbar2.grid(row=0, column=1, padx=10, pady=10, sticky="ns")
-# yscrollbar2.pack(side=RIGHT, fill=Y)
 
 mycanvas2.configure(yscrollcommand=yscrollbar2.set)
 mycanvas2.bind('<Configure>', lambda e: mycanvas2.configure(scrollregion = mycanvas2.bbox('all')))
@@ -92,7 +102,6 @@ mycanvas2.bind('<Configure>', lambda e: mycanvas2.configure(scrollregion = mycan
 
 myframe2 = CTkFrame(mycanvas2, height=850)
 mycanvas2.create_window((0, 0), window=myframe2, anchor="nw")
-
 
 
 def download():
@@ -122,11 +131,15 @@ download_button.grid(row=1, column=0, pady=100, padx=100, sticky="nesw")
 
 
 def getPayload(value):
-    download_button.destroy()
     md2HtmlFile = value.replace(" ", "_")
-    print(md2HtmlFile)
+
     try:
-        if not os.path.exists(f"source/{md2HtmlFile}.html"):
+        if not os.path.exists(f"source/{md2HtmlFile}_README.md"):
+            message = messagebox.showerror("Error", "Download the file first!")
+
+
+        elif not os.path.exists(f"source/{md2HtmlFile}.html"):
+            download_button.destroy() # clears the download/check button
             os.system(f'md2html source/{md2HtmlFile}_README.md > source/{md2HtmlFile}.html')
             with open(f'source/{md2HtmlFile}.html') as f:
                 frame = HtmlFrame(myframe2, horizontal_scrollbar="true", vertical_scrollbar="true")
@@ -134,6 +147,7 @@ def getPayload(value):
                 frame.set_content(x)
                 frame.grid(row=0, column=1, padx=10, pady=10)
         else:
+            download_button.destroy() # clears the download/check button
             with open(f'source/{md2HtmlFile}.html') as f:
                 frame = HtmlFrame(myframe2, horizontal_scrollbar="true", vertical_scrollbar="true")
                 x = f.read()
